requirements.txt
Django~=3.1.4
python-decouple~=3.3
python-gitlab~=2.5.0
django-test-plus~=1.3
